requirements.txt
Django==1.11.11
django-haystack==2.8.1
geetest==3.2.1
jieba==0.39
PyMySQL==0.9.2
pytz==2018.5
qiniu==7.2.2
Whoosh==2.7.4
